myappv2/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.3.0
Django
django-cors-headers==4.2.0
django-phonenumber-field==7.1.0
djangorestframework==3.14.0
wheel
ffmpeg==1.4
geographiclib==2.0
geopy==2.4.0
idna==3.4
numpy==1.25.2
opencv-python==4.6.0.66
phonenumberslite==8.13.21
Pillow
pytz==2023.3
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
urllib3
gunicorn
python-decouple
psycopg2-binary
